chat-bots-profiles/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.5
requests>=2.28.2
aiofiles>=0.8.0
starlette>=0.27.0
openai>=1.1.0
numpy>=1.26.0,<3.0.0
pandas>=1.5.0
matplotlib>=3.5.0
pymupdf>=1.19.0
python-docx>=0.8.11
uuid>=1.30
docx>=0.2.4
PyPDF2>=3.0.0
beautifulsoup4>=4.11.0
markdown>=3.4.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.0
psutil>=5.9.0
chardet>=5.0.0
nltk>=3.8.1
python-pptx>=0.6.21
ebooklib>=0.17.1
lxml>=4.9.0

# Advanced dependencies - uncomment when needed
# langchain>=0.0.267
# langchain-openai>=0.0.2
# chromadb>=0.4.13 